Re-organized logging dir structure + module structure for logging and log handlers + broke apart configure_logging! method + removed StaticallyTaggedLogger but kept the tagging feature via log formatting
