+++ spec/integration/robots_txt_spec.rb
@@ -137,10 +137,10 @@ RSpec.describe 'robots.txt support' do
       end
 
       it 'should abort the redirect and allow all due to robots.txt being treated as missing' do
-        expect_any_instance_of(Crawler::CrawlLogger).to receive(:warn).with(
+        expect_any_instance_of(Crawler::Logging::CrawlLogger).to receive(:warn).with(
           /Purge crawls are not supported for sink type mock. Skipping purge crawl./
         ).at_least(:once).and_call_original
-        expect_any_instance_of(Crawler::CrawlLogger).to receive(:warn).with(
+        expect_any_instance_of(Crawler::Logging::CrawlLogger).to receive(:warn).with(
           /Not following the HTTP redirect.*because the redirect chain is too long/
         ).at_least(:once).and_call_original
